+++ src/main/java/com/sprint/findex/sb02findexteam4/index/info/controller/IndexInfoController.java
@@ -1,0 +1,64 @@
+package com.sprint.findex.sb02findexteam4.index.info.controller;
+
+import com.sprint.findex.sb02findexteam4.index.info.dto.IndexInfoCreateRequest;
+import com.sprint.findex.sb02findexteam4.index.info.dto.IndexInfoDto;
+import com.sprint.findex.sb02findexteam4.index.info.dto.IndexInfoUpdateRequest;
+import com.sprint.findex.sb02findexteam4.index.info.service.IndexInfoService;
+import jakarta.persistence.EntityNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("/api/index-infos")
+public class IndexInfoController {
+
+    private final IndexInfoService indexInfoService;
+
+    // 생성자 주입
+    public IndexInfoController(IndexInfoService indexInfoService) {
+        this.indexInfoService = indexInfoService;
+    }
+
+    @PostMapping
+    public ResponseEntity<IndexInfoDto> createIndexInfo(@RequestBody IndexInfoCreateRequest indexInfoCreateRequest) {
+        try {
+            IndexInfoDto indexInfoDto = indexInfoService.registerIndexInfo(indexInfoCreateRequest);
+            return ResponseEntity.status(HttpStatus.CREATED).body(indexInfoDto); // 201 Created
+        } catch (Exception e) {
+            e.printStackTrace();
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null); // 500 Internal Server Error
+        }
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<IndexInfoDto> findById(@PathVariable Long id) {
+        IndexInfoDto indexInfoDto = indexInfoService.findById(id);
+        return ResponseEntity.status(HttpStatus.OK).body(indexInfoDto);
+    }
+
+
+    @PatchMapping("/{id}")
+    public ResponseEntity<Void> updateIndexInfo(@PathVariable Long id, @RequestBody IndexInfoUpdateRequest indexInfoUpdateRequest) {
+        try {
+            indexInfoService.updateIndexInfo(id, indexInfoUpdateRequest);
+            return ResponseEntity.ok().build(); // 200 OK
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null); // 404 Not Found
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null); // 500 Internal Server Error
+        }
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> delete(@PathVariable long id) {
+        try {
+            indexInfoService.deleteIndexInfo(id);
+            return ResponseEntity.noContent().build(); // 204 No Content
+        } catch (EntityNotFoundException e) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null); // 404 Not Found
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(null); // 500 Internal Server Error
+        }
+    }
+}
