+++ src/main/java/com/sprint/findex/sb02findexteam4/index/info/dto/IndexInfoCreateRequest.java
@@ -1,0 +1,11 @@
+package com.sprint.findex.sb02findexteam4.index.info.dto;
+
+public record IndexInfoCreateRequest(
+   String indexClassification,
+   String indexName,
+   Integer employedItemsCount,
+   String basePointInTime,
+   Double baseIndex,
+   Boolean favorite
+) {
+}
