+++ src/main/java/com/sprint/findex/sb02findexteam4/index/info/dto/IndexInfoSummaryDto.java
@@ -1,0 +1,8 @@
+package com.sprint.findex.sb02findexteam4.index.info.dto;
+
+public record IndexInfoSummaryDto(
+    Long id,
+    String IndexClassification,
+    String IndexName
+) {
+}
