+++ src/main/java/com/sprint/findex/sb02findexteam4/index/info/dto/IndexInfoUpdateRequest.java
@@ -1,0 +1,11 @@
+package com.sprint.findex.sb02findexteam4.index.info.dto;
+
+import java.time.ZonedDateTime;
+
+public record IndexInfoUpdateRequest(
+    Integer employedItemsCount,
+    ZonedDateTime basePointInTime,
+    Double baseIndex,
+    Boolean favorite
+) {
+}
