+++ src/main/java/com/sprint/findex/sb02findexteam4/index/info/entity/SourceType.java
@@ -1,0 +1,6 @@
+package com.sprint.findex.sb02findexteam4.index.info.entity;
+
+public enum SourceType {
+  USER,
+  OPEN_API;
+}
