+++ src/main/java/com/sprint/findex/sb02findexteam4/index/info/service/IndexInfoValidator.java
@@ -1,0 +1,7 @@
+package com.sprint.findex.sb02findexteam4.index.info.service;
+
+import com.sprint.findex.sb02findexteam4.index.info.dto.IndexInfoCreateRequest;
+
+public interface IndexInfoValidator {
+  void validateForCreate(IndexInfoCreateRequest dto);
+}
