--- src/main/java/com/sprint/findex/sb02findexteam4/indexData/service/BasicIndexDataService.java
@@ -1,66 +1,0 @@
-package com.sprint.findex.sb02findexteam4.indexData.service;
-
-import static com.sprint.findex.sb02findexteam4.exception.ErrorCode.INDEX_DATA_ALREADY_EXISTS;
-import static com.sprint.findex.sb02findexteam4.exception.ErrorCode.INDEX_DATA_NOT_FOUND;
-import static com.sprint.findex.sb02findexteam4.exception.ErrorCode.INDEX_INFO_NOT_FOUND;
-
-import com.sprint.findex.sb02findexteam4.exception.AlreadyExistsException;
-import com.sprint.findex.sb02findexteam4.exception.NotFoundException;
-import com.sprint.findex.sb02findexteam4.indexData.dto.IndexDataCreateRequest;
-import com.sprint.findex.sb02findexteam4.indexData.dto.IndexDataResponse;
-import com.sprint.findex.sb02findexteam4.indexData.IndexData;
-import com.sprint.findex.sb02findexteam4.indexData.dto.IndexDataUpdateRequest;
-import com.sprint.findex.sb02findexteam4.indexData.repository.IndexDataRepository;
-import com.sprint.findex.sb02findexteam4.indexInfo.entity.IndexInfo;
-import com.sprint.findex.sb02findexteam4.indexInfo.entity.SourceType;
-import com.sprint.findex.sb02findexteam4.indexInfo.repository.IndexInfoRepository;
-import jakarta.transaction.Transactional;
-import java.time.Instant;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class BasicIndexDataService implements IndexDataService {
-
-  private final IndexDataRepository indexDataRepository;
-  private final IndexInfoRepository indexInfoRepository;
-
-  @Transactional
-  @Override
-  public IndexDataResponse create(IndexDataCreateRequest request, SourceType sourceType) {
-    IndexInfo indexInfo = indexInfoRepository.findById(request.indexInfoId())
-        .orElseThrow(() -> new NotFoundException(INDEX_INFO_NOT_FOUND));
-
-    if (isDuplicated(request.indexInfoId(), request.baseDate())) {
-      throw new AlreadyExistsException(INDEX_DATA_ALREADY_EXISTS);
-    }
-
-    IndexData indexData = IndexData.of(request, indexInfo, sourceType);
-    IndexData createdIndexData = indexDataRepository.save(indexData);
-    return IndexDataResponse.from(createdIndexData);
-  }
-
-  @Transactional
-  @Override
-  public IndexDataResponse update(Long id, IndexDataUpdateRequest request) {
-    IndexData indexData = indexDataRepository.findById(id)
-        .orElseThrow(() -> new NotFoundException(INDEX_DATA_NOT_FOUND));
-
-    IndexData createdIndexData = indexDataRepository.save(indexData.update(request));
-    return IndexDataResponse.from(createdIndexData);
-  }
-
-  @Override
-  public void delete(Long id){
-    IndexData indexData = indexDataRepository.findById(id)
-        .orElseThrow(() -> new NotFoundException(INDEX_DATA_NOT_FOUND));
-
-    indexDataRepository.delete(indexData);
-  }
-
-  @Override
-  public boolean isDuplicated(Long indexInfoId, Instant baseDate) {
-    return indexDataRepository.existsByIndexInfoIdAndBaseDate(indexInfoId, baseDate);
-  }
-}
