--- src/main/java/com/sprint/findex/sb02findexteam4/indexInfo/dto/IndexInfoCreateRequest.java
@@ -1,13 +1,0 @@
-package com.sprint.findex.sb02findexteam4.indexInfo.dto;
-
-import java.time.ZonedDateTime;
-
-public record IndexInfoCreateRequest(
-   String indexClassification,
-   String indexName,
-   Integer employedItemsCount,
-   ZonedDateTime basePointInTime,
-   Double baseIndex,
-   Boolean favorite
-) {
-}
