--- src/main/java/com/sprint/findex/sb02findexteam4/indexInfo/dto/IndexInfoSummaryDto.java
@@ -1,8 +1,0 @@
-package com.sprint.findex.sb02findexteam4.indexInfo.dto;
-
-public record IndexInfoSummaryDto(
-    Long id,
-    String IndexClassification,
-    String IndexName
-) {
-}
