--- src/main/java/com/sprint/findex/sb02findexteam4/indexInfo/entity/SourceType.java
@@ -1,6 +1,0 @@
-package com.sprint.findex.sb02findexteam4.indexInfo.entity;
-
-public enum SourceType {
-  USER,
-  OPEN_API;
-}
