--- src/main/java/com/sprint/findex/sb02findexteam4/indexInfo/mapper/IndexInfoMapper.java
@@ -1,22 +1,0 @@
-package com.sprint.findex.sb02findexteam4.indexInfo.mapper;
-
-import com.sprint.findex.sb02findexteam4.indexInfo.dto.IndexInfoDto;
-import com.sprint.findex.sb02findexteam4.indexInfo.entity.IndexInfo;
-import org.springframework.stereotype.Component;
-
-@Component
-public class IndexInfoMapper {
-
-  public IndexInfoDto toDto(IndexInfo entity) {
-    return new IndexInfoDto(
-        entity.getId(),
-        entity.getIndexClassification(),
-        entity.getIndexName(),
-        entity.getEmployedItemsCount(),
-        entity.getBasePointInTime(),
-        entity.getBaseIndex(),
-        entity.getSourceType(),
-        entity.getFavorite()
-    );
-  }
-}
