--- src/main/java/com/sprint/findex/sb02findexteam4/indexInfo/service/IndexInfoValidator.java
@@ -1,7 +1,0 @@
-package com.sprint.findex.sb02findexteam4.indexInfo.service;
-
-import com.sprint.findex.sb02findexteam4.indexInfo.dto.IndexInfoCreateRequest;
-
-public interface IndexInfoValidator {
-  void validateForCreate(IndexInfoCreateRequest dto);
-}
