+++ src/main/java/com/sprint/findex/sb02findexteam4/sync/entity/SyncJobHistory.java
@@ -1,6 +1,6 @@
 package com.sprint.findex.sb02findexteam4.sync.entity;
 
-import com.sprint.findex.sb02findexteam4.indexInfo.entity.IndexInfo;
+import com.sprint.findex.sb02findexteam4.index.info.entity.IndexInfo;
 import jakarta.persistence.*;
 import jdk.jfr.Timestamp;
 import lombok.Getter;
