+++ src/main/java/com/sprint/findex/sb02findexteam4/sync/service/AutoSyncConfigService.java
@@ -1,6 +1,6 @@
 package com.sprint.findex.sb02findexteam4.sync.service;
 
-import com.sprint.findex.sb02findexteam4.indexInfo.entity.IndexInfo;
+import com.sprint.findex.sb02findexteam4.index.info.entity.IndexInfo;
 import com.sprint.findex.sb02findexteam4.sync.dto.AutoSyncConfigDto;
 import com.sprint.findex.sb02findexteam4.sync.dto.AutoSyncConfigUpdateCommand;
